Added a few attributes to get filtering up and running.
The filter is very important for helping users access the runewords.
Angulars built in features are great, but an advanced filter is needed
so sort through runes, properties, or item types only. I think this is
where the real fun of this app will happen.

Next three big steps to get this app in a good place:
1. Add advanced, pragmatic filtering a Diablo 2 player will use when they
alt tab out of the game.
2. Add images (to the API) and host them so the site can start looking good.
Overall just turn it up when it comes to the style.
3. Clickable everything. Something good about the d2 sites which do exist is
that they have wiki-like links to everything. If I add a link to each property
, rune, and item type, but without the 2 second load those old sites come with,
the site will bring actual value and usability to users.