Added a dropdown to filter item types.
I wanted to use a custom filter, but it was complete overkill for the simplicity
of the drop down filtering. I noticed in making a new factory I could make use of
constants for the api's base url. I'm going to priortize filtering by property
categories next, however, since that type of feature is the true intent of this app.