straightened up HTML.
I decided with all the tags angular brings in it's best to organize your HTML
in a way which best spares the reader. In this case, I added a new line for each
tag after the first on each element with more than one tag. I also separated out
the identifier tags (ie: <p>) from the content of the tags. This already looks a
lot less messy, and I know just where the checkboxes are to go.